Move `TableClickBehaviorView` to its own directory
