+++ frontend/src/metabase/dashboard/components/ClickBehaviorSidebar/TableClickBehaviorView/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./TableClickBehaviorView";
